% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployrandomForest.R
\name{create_web_api}
\alias{create_web_api}
\title{operationalization: SERVER side}
\usage{
create_web_api()
}
\value{
A locally REST Api gateway.
}
\description{
Takes in a trained model and changes it to JSON by using the plumber, jsonlite and randomForest-packages.
}
