% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployrandomForestRUN.R
\name{start_api_plumber}
\alias{start_api_plumber}
\title{operationalization: SERVER side RUN}
\usage{
start_api_plumber(x)
}
\arguments{
\item{x}{A numeric value which defines the port. Usally one should use x=8000.}
}
\value{
Start a locally REST Api gateway.
}
\description{
Start a locally REST Api gateway by using plumber. Usally one should use port=8000.
}
